[git] Fix Git ignore of dummy files
In the case of regression directories, we don't want to ignore the whole
`tmp` and `tmp_enc` dir. We want to include them because we use them during
tests. To accomplish that, we have a .dummy file inside. The proper way
of doing it is to ignore all files and directories under `tmp` and
`tmp_enc`. Next exclude .dummy ones from ignoring list.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>