[LibOS] Handle `IP_RECVERR` socket option as a no-op
This socket option allows for later retrieval of detailed error
descriptions that happened on the socket. We do not support such
retrieval, so allowing to set this option basically defers the moment
the app notices that it's not supported from the call to `setsockopt`,
to the moment the error condition actually happens (which might be
never).

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>