[common] Drop offset from the protected file structure
The `offset` field supposed to represent the current file position
(user's view). This isn't true. The actual offset of the file is kept in
the LibOS, not in the protected files. We can observe this when lseek(2)
is used. Also there isn't any API that allows us to get a current offset
from protected file structure. The offset is changed when the LibOS wants
to read/write the data.

This field is ambiguous, so it's better to remove it and pass the
requested offset to each function.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>