[LibOS] Override `argv[0]` with file pathname for interpreter scripts
In Linux, if the pathname argument of `execve()` specifies an
interpreter script, then interpreter will be invoked with the following
arguments:

    interpreter [optional-arg] pathname arg...

where `pathname` is the pathname of the file specified as the first
argument of `execve()`, and `arg...` is the series of words pointed to
by the `argv` argument of `execve()`, starting at `argv[1]`.

The above text is taken verbatim from the man page of `execve`.

Signed-off-by: aneessahib <anees.a.sahib@intel.com>