[LibOS] Print debug message in `load_and_check_shebang()`
Print debug message on `-ENOEXEC` error (failure to read the shebang
line) instead of the error message. The error code is already propagated
to the app that called `execve()`, and sometimes apps open "wrong" files
(e.g., LTP does it), so it is not a Gramine error and must be silenced.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>