[LibOS] Strip trailing spaces/tabs in the shebang line
Previously, the "load shebang" logic didn't strip trailing spaces/tabs,
which resulted in additional empty arguments to the interpreter (e.g.,
`#!/bin/bash ` was interpreted as "start `/bin/bash` with the first
argument ` `" -- and Bash failed because it expected the first argument
to be the program to run). This commit also adds handling of tabs in
addition to spaces.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>