[CI] Add a VM-based Jenkins pipeline
New Jenkins pipeline runs a Docker container that creates a minimal VM
based on https://github.com/gramineproject/device-testing-tools repo and
runs a subset of Gramine tests, in particular, the device IOCTL tests.
The pipeline uses Ubuntu 22.04 with modern Linux kernel (to have an
upstream SGX driver and support for SGX in KVM) and QEMU/KVM to run the
VM.

A LibOS regression test `device_ioctl` is added to test the CI;
currently it is minimal but will be expanded when ioctl passthrough
functionality is added to Gramine.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>