[LibOS] Add support for `setsid()`/`getsid()`
Previously, session-management syscalls e.g., `setsid()`, `getsid()`,
were not supported in Gramine. However, they are required for many
workloads, particularly the ones that use JVM.

This commit adds basic support for session-management syscalls and
enables the related tests. As a side effect of the introduction of
`setsid()`, `pgid` and `sid` in struct `libos_process` are now both not
suitable for atomic access so we introduced a new global lock and used
it everywhere where we access `pgid` and `sid`.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>