[LibOS,PAL] Treat (e)poll error and hang-up events separately
Previously, `(E)POLLRDHUP` was not handled and `(E)POLLHUP` was always
reported together with `(E)POLLERR` due to the limitation of PAL API,
which did not distinguish these conditions. This may cause some
workloads to ignore the fact that the socket was closed by the peer and
hence keep it in the `CLOSE-WAIT` state.

This commit:
* splits (e)poll error events from hang-up events;
* adds support for `(E)POLLRDHUP` but treats/reports it as `(E)POLLHUP`,
  i.e., not distinguishing these two states.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>