[PAL/Linux] Log a message on `munmap()` returning `-ENOMEM`
Linux can return `-ENOMEM` if a request to `munmap()` will exceed the
process's maximum number of mappings (this is possible in case of
partial memory unmap, when Linux would have to create two mappings out
of one). Gramine's LibOS is not ready for such failures, and also it is
specific to the Linux PAL, so this commit simply makes this failure more
visible and provides a hint for the user what can be done on the host.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>