[PAL/VM,TDX] Account for int wrap around in virtio queues
Previously, virtio drivers assumed that the host VMM is malicious if it
would return a `used_idx` in some virtio queue that is less than the
VM-memoized last seen `used_idx`. However, this case is possible on int
wrap around. Thus, this commit modifies the check to verify that the
host-returned index is at most queue-size items ahead of the VM-memoized
index, accounting for int wrap.

Note that this check concerns only console and vsock drivers. The fs
driver is not problematic because it always assumes an increment of 1 in
the virtio RX queue (this increment accounts for possible int wrap).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>