[PAL/VM,TDX] Start from VMM-provided UNIX timestamp
Previously, Gramine-VM and Gramine-TDX did not have a correct UNIX time
on startup. They simply started with the timestamp of `0` (which in
UNIX-time format means "1. January 1970").

This commit gets the UNIX time value on startup from the VMM using
"FW CFG" feature of QEMU. The obtained time value is untrusted so checks
are added to guarantee it's in a reasonable range. In the future, it
should be obtained from a trusted time source, like a trusted remote
server.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>