[PAL/VM,TDX] Remove timeout objects via `deregister_timeout()`
Previously, wait-with-timeout functions registered a timeout (in a
global list) using `register_timeout()`. This func allocated a timeout
object and added it to the list of pending timeouts. Another func
`notify_about_timeouts_uninterruptable()` would remove the triggered
timeout from the list and free the timeout object itself.

However, if the wait-with-timeout function finishes before the timeout
is triggered, then the timeout would become "dangling". In reality, when
the wait-with-timeout function finishes, the timeout must be removed
from the list and freed (i.e. by the function itself).

This commit forces wait-with-timeout functions to take ownership of the
timeout object, returned by `register_timeout()` and deregister this
timeout upon finishing via `deregister_timeout()`. As a side effect, the
two users of timeout -- `PalEventWait()` and `PalStreamsWaitEvents()` --
are refactored to have uniform timeout-handling code.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>