[PAL/TDX] Extend RTMR2 with manifest file contents
Previously, RTMR2 was erroneously extended with `loader.entrypoint`
(i.e. the LibOS binary). This was (a) insecure because we were reading
this file untrusted and unchecked so a malicious host could replace this
file later, and (b) irrelevant because the manifest file -- the actual
root of trust -- was not reflected anywhere in TDX measurements.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>