[PAL/vm-common] Use array and hash table for vsock connections
Previously in Gramine-VM and Gramine-TDX, open connections in the vsock
(virtual socket) driver were kept in a static array of max 10
connections. This had two drawbacks:
- Max number of connections was limited to 10. Even if we would increase
  this limit, unnecessary memory needs to be pre-allocated.
- Code had to iterate over the whole connections array to find a
  corresponding connection, which was not efficient.

This commit uses an ever-growing array to create a mapping between a
file descriptor (FD) and a given vsock connection. By this, finding a
connection based on the FD is an O(1) operation (frequent on
socket-related syscalls). Since we completely control FD numbers, we try
to keep the FDs at small values, so that the array doesn't grow too
large (by reusing the FDs).

This commit uses a hashtable with `virtio_vsock_connection::host_port`
as the key, such that finding a connection based on the host port is an
O(1) operation (frequent on receival of a packet from the host).

Signed-off-by: Kailun Qin <kailun.qin@intel.com>