[LibOS] Add parser for `flock()` syscall
This is a preparatory commit to introduce flock (BSD locks) support.
Note that for the parser to work, the `flock()` syscall must be
implemented in LibOS. Currently, the implementation just returns
`-ENOSYS` (i.e., no functional change), except for the corner case of
deprecated `LOCK_MAND` (`LOCK_MAND` mode is nowadays ignored on Linux
and always succeeds, so it is easy to add this case).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>