[common] Do not expand `__FILE_NAME__` to our func call
Previously we had a function that emulated `__FILE_NAME__` macro (which
was introduced only recently in GCC12 and Clang9) by truncating the file
name in a loop. It led to unnecessary calls to this func even when logs
were disabled, ultimately resulting in perf overhead. We observed around
5% overhead because of this on file-I/O intensive benchmarks (because
there we print a lot of log_debug info).

This commit fixes this on non-debug builds (i.e. `debugoptimized` and
`release`), by using a macro that has a cascade of IFs and is optimized
at compile-time, thus incurring no perf overhead.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>