[PAL/vm-common] Revert optimization of lazy memory allocation
Commit 29bce8c introduced lazy memory allocation, where newly-allocated
pages are NOT memset to zero immediately, but lazily on a #PF. This
turned out to be a case of premature optimization. It only complicated
things, and made the #PF handler unnecessarily complex.

The particular issue that led to this revert commit: in multi-core
environments, a #PF handler must send an IPI to update PTEs in other
CPUs (TLB Shootdown). This is hard, slow, and missing this IPI led to a
complex data race.

Note that not the whole commit 29bce8c is reverted -- the side effect of
"interrupt stack table" entries in TSS is useful and thus stays.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>