[PAL/vm-common] vsock: Correctly return "write possible" in poll
Previously, the "write possible" event for sockets during
select/poll/epoll was always reported. This is not true when there are
currently no available slots for writing in the TX virtqueue in the
vsock driver. This simplification led e.g. Memcached to incorrect
behavior, trying to send packets on a full TX queue.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>