[PAL/vm-common] Protect the list of pending timeouts in interrupt context
Now that Gramine-TDX supports multi-core environments, it is possible to
have CPU1 modifying the list of pending timeouts via
`register_timeout()` or `deregister_timeout()` while CPU0 at the same
time operates on this list in `notify_about_timeouts_uninterruptable()`.
So we need to add a spinlock around this code block.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>