[PAL/Linux-SGX] Log invalid search domains/hostname in `post_callback()`
Gramine's logging configuration is done in the PAL-common logic (to
avoid duplication), which is executed after the PAL-specific logic. So
previously, the warnings about invalid search domain names and/or
hostname (which are checked and produced during the PAL-specific logic)
didn't show up in Gramine logs.

This commit saves the invalid DNS information, and prints the warnings
in `post_callback()`, which is after logging is configured.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>