[PAL/vm-common] console: correctly reuse consumed descriptors in tq
This commit introduces correct reuse of descriptors during console
writes. Previously our virtio-console implementation did not verify the
consumption of descriptors on the transmit queue; instead it used
optimistic garbage collection (assuming that if the shared messages
buffer is full, the host consumed the previous descriptors and we can
remove them). This was not safe, and could lead to jumbled output. In
this commit, we do the correct thing: examine the `tq->used` array.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>