[tools] Introduce Gramine VM ID in `gramine-vm.in`
Previously, we hard-coded the `guest-cid` (CID) in virtio-vsock to `10`
and the UNIX socket path for `vhostfs` to `/tmp/vhostfs`. This made
running two instances of `gramine-vm` or `gramine-tdx` at the same time
impossible, with errors like these:
- from `virtio-vsock`: `unable to set guest cid: Address already in use`
- from `virtiofsd`: `Error creating pid file '/tmp/vhostfs.pid'`

This commit introduces the `GRAMINE_VM_ID` variable used to configure
the `guest-cid` and the path for `virtiofs`. `GRAMINE_VM_ID` is
determined based on the already used `guest-cid`s. We scan the running
QEMU processes and identify the `guest-cid`s of the running VMs.
Starting from our default value (`10`), we search for the next available
`guest-cid`. In this way, we avoid collisions between the instances.

Signed-off-by: Dimitrios Stavrakakis <dimitrios.stavrakakis@intel.com>