[PAL/vm-common] Use dummy TCB with stack canary on boot/termination
Previously, during boot of the very first (temporary) thread and during
termination of threads, the GS register was set to zero. Recall that GS
register is supposed to point to the TCB (Threat Control Block) in
Gramine, in particular `gs:0x8` should contain the stack canary used by
GCC's stack protector. So previously we had a bug of trying to read the
canary from address `0x0008`, which is accessible only by accident
(Gramine's common VM code maps all pages including the zero page as
Present and RWX).

This bug was uncovered when implementing proper page permissions for
VM-based PALs. This commit replaces a zero GS register with a pointer to
a dummy TCB that contains a hard-coded (and always accessible) stack
canary.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>