[PAL/vm-common] vsock: drain TX virtio queue if it is full
Previously, when trying to send a normal packet on vsock TX queue that
is full, we bailed out immediately and dropped this packet. This commit
adds a retry -- if the queue is full, then we drain it first, then try
to send the packet again, and only after the second failed attempt we
bail out (and print a warning that we dropped the packet).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>