[PAL/vm-common] Set write permission on file-backed mmap
Previously, we set whatever permission the user requested on a
file-backed mmap immediately. This is wrong: the subsequent code will
update the mmapped region with file contents, so first the region must
be mmapped with write permission too, and only after the update the
write permission must be revoked (if read-only was requested by user).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>