[PAL/vm-common] Correctly set up page tables for non-GB-aligned VM sizes
Previously, there was a bug in `pagetables_init()`: PDP entries may be
"forgotten" for VMs with sizes not aligned to 1GB. For example, a VM
with 1.5GB would generate two Page Directory Tables, but there will be
only one PDP entry to the first of these two PDTs.

This commit fixes this: we use the ceiling function instead of floor
function during calculation of page tables' entries.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>