[PAL/vm-common] Periodically trigger `g_streams_waiting_events_futex`
Current futex triggering has a data race: if the thread that is supposed
to wait on poll/epoll (emulated in PAL via `PalStreamsWaitEvents()`) is
currently running, and another thread (on another CPU) triggers the
polled-on futex, this event is lost. This is because all polls wait on a
single global `g_streams_waiting_events_futex`, so if two threads (e.g.
Async and IPC threads) poll at the same time, then only one of them
notices the wake up, and the futex is considered triggered and not
reflected in another (currently running) thread on the next poll wait.
This may lead to hangs, e.g., in the IPC thread that waits without any
timeout.

We work around this peculiarity of current `PalStreamsWaitEvents()`
implementation by periodically triggerring this global futex, in
particular, during timer interrupts on CPU0. A proper fix would be to
introduce a triggering mechanism that uses fine-grained futexes.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>