[LibOS] Wake up thread that can handle SIGTERM on host-injected SIGTERM
Some apps (e.g. MongoDB) create a separate thread that has a single job
-- to wait for a SIGTERM signal and then perform graceful app
termination. This special thread issues `sigtimedwait(SIGTERM)` and
waits forever for this signal. Other threads instead block this signal
(add it to their sigmask), so that they don't need to care about
reacting to this signal.

Previously, our LibOS would try to handle SIGTERM on the thread that got
chosen by the host Linux (the thread is arbitrarily chosen). But with
high probability, on such apps this thread will be the not-special-one,
and the app would effectively ignore SIGTERM.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>