[tests] Allow POSIX shared memory objects in LTP manifest template
Due to the lack of POSIX shared memory support, Gramine previously had
an LTP configuration under SGX specifically for skipping those tests
that require it.

Now that Gramine has added shared untrusted memory support and it's fine
to use it in functionality tests, in this commit we mount the `/dev/shm`
directory and allow files within it in the manifest to unblock the
relevant LTP tests.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>