[common] Add explicit fsync callback to Protected Files
Previously, Protected Files didn't have a way to ensure that the data is
physically stored on disk (aka fsync). This is in contrast to the "file
flush" operation, which only flushes user-space buffers to the OS kernel
buffers. (Gramine confusingly uses the term "flush" to mean flush and
fsync, depending on the context.)

This explicit fsync is important in e.g. persistent database workloads.
It was detected on a RocksDB workload.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>