[glibc] Fix redefinition of `_FORTIFY_SOURCE`
The new Ubuntu release, Noble, includes a version of GCC with a spec that
defines `_FORTIFY_SOURCE` when optimization is enabled:
```
$ gcc -dumpspecs | grep _FORTIFY_SOURCE
 %{!O0:%{O*:%{!D_FORTIFY_SOURCE=*:%{!U_FORTIFY_SOURCE:-D_FORTIFY_SOURCE=3}}}
```

However, some files in glibc need to be built without `_FORTIFY_SOURCE`,
such as `vprintf.c`. The patch from Ubuntu Noble changes the method of
undefining _FORTIFY_SOURCE to -U_FORTIFY_SOURCE. Previously it was limited
only to the preprocessor (-Wp,-U_FORTIFY_SOURCE). The new method also
undefines the definition from the spec.
The patch addresses some other non-standard redefinitions of
_FORTIFY_SOURCE as well.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>