[PAL/vm-common] console: use `virtq_used_elem::len` to learn RQ packet size
Previously, there was a bug in virtio-console's receive queue (RQ): our
code read the buffer size specified in the descriptor, thinking this is
the size of the incoming packet (i.e., number of bytes written by the
host into the buffer). Instead, the size of the incoming packet must be
read from the user ring, i.e. from `virtq_used_elem::len`.

With this fix, there is no need for the "search for NUL symbols"
workaround (which was also incorrect and worked by accident).

Additionally, this commit removes the re-initialization of
descriptors, as they are read-only for the host and never changed by our
driver. Note that we already verify values in descriptors (in
particular, the address), so if malicious host modifies the descriptors,
it will be detected.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>