[tools] Disable char backend multiplexer in `gramine-vm.in`
Previously, QEMU's stdio was multiplexed on several serial devices: the
QEMU monitor, the `-nographic` device, and virtio-console. For some
reason (wasn't able to root cause), this multiplexer leads to "jumps" in
stdin input when using Gramine's virtio-console implementation. This
behavior is observed on newer QEMU (e.g. v8.2.2) and doesn't appear on
older versions. This may be a subtle bug in QEMU. Regardless, there is
no reason to use the multiplexer, so it is removed.

Also, `-no-hpet` is deprecated in newer QEMU, so this commit uses
`-machine hpet=off` instead.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>