[PAL/vm-common] Fix undefined behavior in `find_fw_cfg_selector()`
Previously, `find_fw_cfg_selector()` had a UB: the main variable
`uint32_t fw_cfg_files_count` was aliased by the 4-bytes array
`uint8_t fw_cfg_files_count_raw[4]`. This was done for ease of port I/O,
since only byte-sized port I/O is used for FW_CFG device. The hope was
that updating bytes in `fw_cfg_files_count_raw` would transparently
update the main variable. However, this is UB, and GCC build with
optimizations (i.e., in the release mode) assumed that the main variable
`fw_cfg_files_count` is not updated and is still zero, which led to
incorrect result of the function, which ultimately led to abnormal
termination of the VM in the early stages.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>