[PAL/Linux-SGX] Rename `sgx_entry` to `sgx_do_host_ocall`
The file `host_entry.S` contains low-level code that executes
ECALL/OCALL/AEX flows in the untrusted runtime of Gramine. Previously,
this file contained the label `sgx_entry` that has a misleading name --
it doesn't "enter the SGX enclave" but instead it "performs an OCALL on
the host", i.e. executes the syscall on the host. This commit renames
this label to avoid confusion. Also, some explanations are added.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>