[LibOS] Make async worker persistent
Previously, the async worker thread was spawned on-demand and terminated
after an idle timeout. This is entirely unnecessary — the cost of having
a persistent async thread is near-0, and it simplifies the logic.

Also fixes a long-standing issue about not waiting for the async thread
to exit (https://github.com/gramineproject/graphene/issues/440).

Signed-off-by: Marcelina Kościelnicka <mwk@0x04.net>