[Docs] Change /usr/share/keyrings to /etc/apt/keyrings
Since apt 2.4, the distro provides /etc/apt/keyrings for user-managed
package signing keys and /usr/share/keyrings is recommended for use with
keys managed by packages.

Debian 11 has older apt, so the directory is not present in the system.
Therefore it is left as is, instead of adding `mkdir -p`.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>