[tests] Add `-Wno-implicit-function-declaration` for LTP build
Currently used version of LTP has a bug that results in the following
build error, e.g. when using Clang v18 on Ubuntu 24.04:

    clone301.c:136:8: error: call to undeclared function 'pidfd_send_signal'

This is because Clang promoted "implicit function declaration" check
from a warning to an error starting from Clang v16. This commit adds
`-Wno-implicit-function-declaration` flag to CFLAGS during LTP build.

Newer LTP releases fixed this bug, so in the future, when Gramine uses
newer LTP, this commit should be reverted.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>