[LibOS/tests] Fix openmp manifest to add shm tmpfs
There is a known issue in some versions of `libomp` (the
LLVM implementation of OpenMP), and probably also in `libiomp` (Intel
implementation of OpenMP). Those versions require /dev/shm to
exist. This commit modifies the openmp manifest to add this dir as a
tmpfs mount. With this fix, the files under /dev/shm/ are not visible on
the host, so security of the application remains intact.

Based on the suggestion by Dmitrii Kuvaiskii. See
https://github.com/gramineproject/gramine/discussions/1780#discussioncomment-8589083

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>