[PAL/Linux] Log a message on `mmap()` returning `-ENOMEM`
Linux can return `-ENOMEM` if a request to `mmap()` exceeds the
process's maximum number of mappings. After commit f7463c5
"[LibOS,PAL/Linux-SGX] Add EDMM lazy allocation support", this can occur
during the checkpoint/restore on the Linux PAL. In such case, each
committed page is sent separately and Linux creates a mapping for each
one, which can result in a large number of mappings.

This commit makes this failure more visible and provides a hint to the
user on what can be done on the host.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>