[LibOS] Fix memmgr free list performance bug
The `CHECK_LIST_HEAD` macro in `common/include/list.h` is used in memmgr
to traverse the free list and perform assertion checks. However, the
loop is not removed by the compiler even in release mode, leading to
unnecessary overhead. This commit fixes this by changing the macro to
`(void)0` in release mode.

Signed-off-by: Adrian Lutsch <adrian.lutsch@cs.tu-darmstadt.de>