[meson,packaging] Bump meson dependency to >= 0.58
The problem is, str.replace was introduced in commit 31dbd9b24f5b to
adjust the version to be acceptable by Python tooling, but that function
is available only from meson 0.58. Currently we can't use any meson
version newer than 0.58, because this is the latest version available in
AlmaLinux 8 repositories, which we use to build *EL packages.

Fixes: 31dbd9b24f5b ("[python] Fix version in distribution")
Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>